// src/app/api/roboticsAudit/audit-program.ts
export interface AuditProgram {
    audit_program_id: number;
    ap_name: string;
    ap_desc: string;
    audit_id: number;
    au_level_3_id: number;
    au_level_3_desc: string;
    ap_schedule_date: string;
    ap_schedule_time: number;
    frequency_id: number;
    next_run: string;
    last_run: string;
    audit_uid: string;
    audit_program_uid: string;
    au_level_3_uid: string;
    full_name: string;
    schedule_status: boolean;
    schedule_enable: boolean;
    audit_board_id: string;
    audit_name: string;
    department_id: number;
    banner_id: number;
    au_level_2_id: number;
    review_year: number;
    quarter: number;
    start_date: string;
    end_date: string;
    acp_audit: boolean;
    user_access_id: string;
    audit_schedule?: any;
    results_url?: any;
    last_run_date: string;
    au_level_2_uid?: any;
    au_level_2_desc?: any;
    department_uid?: any;
    banner?: any;
    created_date: string;
    created_by: number;
    updated_date: string;
    updated_by: number;
    useracsname?: any;
    organization_id: string;
    organization?: any;
}

export interface Audit {
    audit_id: number;
    audit_uid: string;
    audit_board_id: string;
    audit_name: string;
    department_id: number;
    banner_id: number;
    au_level_2_id: number;
    review_year: number;
    quarter: number;
    start_date: string;
    end_date: string;
    acp_audit: boolean;
    user_access_id: string;
    audit_schedule?: any;
    results_url?: any;
    last_run_date: string;
    full_name: string;
    au_level_2_uid: string;
    au_level_2_desc: string;
    department_uid: string;
    banner?: any;
    created_date: string;
    created_by: number;
    updated_date: string;
    updated_by: number;
    useracsname: string;
    organization_id: string;
    organization: string;
}